bugfix: bus/device manager对卸载逻辑的处理错误 (#385)
* 移动位置

* bugfix: bus/device manager对卸载逻辑的处理错误
