+++ kernel/src/driver/base/device/driver.rs
@@ -119,6 +119,6 @@ pub fn driver_register(driver: Arc<dyn Driver>) -> Result<(), DriverError> {
 /// @return: 操作成功，返回()，操作失败，返回错误码
 #[allow(dead_code)]
 pub fn driver_unregister(driver: Arc<dyn Driver>) -> Result<(), DriverError> {
-    DRIVER_MANAGER.add_driver(driver.id_table(), driver);
+    DRIVER_MANAGER.remove_driver(&driver.id_table());
     return Ok(());
 }
